+++ bugsnag/configuration.py
@@ -20,12 +20,16 @@ from bugsnag.middleware import (
     MiddlewareStack,
     skip_bugsnag_middleware
 )
-from bugsnag.utils import (fully_qualified_class_name,
-                           partly_qualified_class_name,
-                           validate_str_setter, validate_bool_setter,
-                           validate_iterable_setter,
-                           validate_required_str_setter,
-                           validate_int_setter)
+from bugsnag.utils import (
+    fully_qualified_class_name,
+    partly_qualified_class_name,
+    validate_str_setter,
+    validate_bool_setter,
+    validate_iterable_setter,
+    validate_required_str_setter,
+    validate_int_setter,
+    validate_path_setter
+)
 from bugsnag.delivery import (create_default_delivery, DEFAULT_ENDPOINT,
                               DEFAULT_SESSIONS_ENDPOINT)
 from bugsnag.uwsgi import warn_if_running_uwsgi_without_threads
@@ -37,6 +41,14 @@ try:
 except ImportError:
     from bugsnag.utils import ThreadContextVar
     _request_info = ThreadContextVar('bugsnag-request', default=None)  # type: ignore  # noqa: E501
+
+
+try:
+    from os import PathLike
+except ImportError:
+    # PathLike was added in Python 3.6 so fallback to PurePath on Python 3.5 as
+    # all builtin Path objects inherit from PurePath
+    from pathlib import PurePath as PathLike  # type: ignore
 
 
 __all__ = ('Configuration', 'RequestConfiguration')
@@ -365,9 +377,9 @@ class Configuration:
         return self._project_root
 
     @project_root.setter  # type: ignore
-    @validate_str_setter
-    def project_root(self, value: str):
-        self._project_root = value
+    @validate_path_setter
+    def project_root(self, value: Union[str, PathLike]):
+        self._project_root = str(value)
 
     @property
     def proxy_host(self):